Moved the simulation example to a sub folder
