+++ example/simulating_callbacks/flashing_the_LED.py
@@ -1,7 +1,7 @@
 import tkinter as tk
 
-from mychip.reg_model.mychip import mychip_cls
-from mychip.peakrdl_python import CallbackSet
+from example.simulating_callbacks.mychip.reg_model.mychip import mychip_cls
+from example.simulating_callbacks.mychip.peakrdl_python import CallbackSet
 
 class ChipSim:
